iprt/critsect.h: Shut up sign conversion warnings.  bugref:9790
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@85226 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/critsect.h
@@ -336,7 +336,7 @@ DECLINLINE(bool) RTCritSectIsInitialized(PCRTCRITSECT pCritSect)
  */
 DECLINLINE(uint32_t) RTCritSectGetRecursion(PCRTCRITSECT pCritSect)
 {
-    return pCritSect->cNestings;
+    return (uint32_t)pCritSect->cNestings;
 }
 
 /**
